requirements.txt
astroid==3.3.11
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.6
cryptography==46.0.1
dill==0.4.0
Flask==3.1.2
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
platformdirs==4.4.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.8
pytest==8.4.2
pytest-cov==7.0.0
requests==2.32.5
tomlkit==0.13.3
urllib3==2.5.0
Werkzeug==3.1.3
